requirements.txt
beautifulsoup4
requests
dnspython
pyopenssl
ndg-httpsclient
pyasn1
